requirements.txt
pynput
ffmpeg-python
PySide6
mss
opencv-python
google-genai
python-dotenv
prisma
asyncpg 